+++ {{cookiecutter.package_name}}/{{cookiecutter.app_name}}/cms_plugins.py
@@ -1,0 +1,9 @@
+"""
+CMS Plugins installations
+"""
+from cms.plugin_pool import plugin_pool
+
+from .plugins.blog import BlogPlugin
+
+
+plugin_pool.register_plugin(BlogPlugin)