+++ {{cookiecutter.package_name}}/{{cookiecutter.app_name}}/factories/__init__.py
@@ -1,10 +1,11 @@
-from .blog import BlogFactory
+from .blog import BlogFactory{% if cookiecutter.include_cmsplugin %}, BlogPluginModelFactory{% endif %}
 from .article import ArticleFactory
 from .user import UserFactory
 
 
 __all__ = [
     "ArticleFactory",
-    "BlogFactory",
+    "BlogFactory",{% if cookiecutter.include_cmsplugin %}
+    "BlogPluginModelFactory",{% endif %}
     "UserFactory",
 ]