+++ {{cookiecutter.package_name}}/{{cookiecutter.app_name}}/factories/blog.py
@@ -1,6 +1,6 @@
 import factory
 
-from ..models import Blog
+from ..models import Blog{% if cookiecutter.include_cmsplugin %}, BlogPluginModel{% endif %}
 
 
 class BlogFactory(factory.django.DjangoModelFactory):
@@ -11,3 +11,15 @@ class BlogFactory(factory.django.DjangoModelFactory):
 
     class Meta:
         model = Blog
+
+{% if cookiecutter.include_cmsplugin %}
+class BlogPluginModelFactory(factory.django.DjangoModelFactory):
+    """
+    Factory to create instance of a BlogPluginModel.
+    """
+    blog = factory.SubFactory(BlogFactory)
+    limit = factory.Faker("random_int", min=1, max=5)
+
+    class Meta:
+        model = BlogPluginModel
+{% endif %}