+++ {{cookiecutter.package_name}}/{{cookiecutter.app_name}}/forms/__init__.py
@@ -1,0 +1,6 @@
+from .blog import BlogPluginForm
+
+
+__all__ = [
+    "BlogPluginForm",
+]