+++ {{cookiecutter.package_name}}/{{cookiecutter.app_name}}/forms/blog.py
@@ -1,0 +1,13 @@
+from django import forms
+
+from ..models.blog import BlogPluginModel
+
+
+class BlogPluginForm(forms.ModelForm):
+    class Meta:
+        model = BlogPluginModel
+        exclude = []
+        fields = [
+            "blog",
+            "limit",
+        ]