+++ {{cookiecutter.package_name}}/{{cookiecutter.app_name}}/models/__init__.py
@@ -1,8 +1,9 @@
-from .blog import Blog
+from .blog import Blog{% if cookiecutter.include_cmsplugin %}, BlogPluginModel{% endif %}
 from .article import Article
 
 
 __all__ = [
-    "Blog",
+    "Blog",{% if cookiecutter.include_cmsplugin %}
+    "BlogPluginModel",{% endif %}
     "Article",
 ]